refac: 関数の切り出し
